Cast to output, not input in numba dispatch of scalar Softplus
